+++ runtime/engine/src/test/java/io/aklivity/zilla/runtime/engine/internal/EngineMetricsIT.java
@@ -104,8 +104,7 @@ public class EngineMetricsIT
         long[][] gaugeIds = engine.gaugeIds();
 
         // THEN
-        // gaugeIds[0] is coming from test.gauge in server.yaml
-        assertThat(gaugeIds[2], equalTo(new long[]{3L, 7L}));
+        assertThat(gaugeIds[gaugeIds.length - 1], equalTo(new long[]{3L, 7L}));
     }
 
     @Test