+++ runtime/engine/src/test/java/io/aklivity/zilla/runtime/engine/test/EngineRule.java
@@ -23,6 +23,7 @@ import static io.aklivity.zilla.runtime.engine.EngineConfiguration.ENGINE_ROUTED
 import static io.aklivity.zilla.runtime.engine.EngineConfiguration.ENGINE_SYNTHETIC_ABORT;
 import static io.aklivity.zilla.runtime.engine.EngineConfiguration.ENGINE_WORKERS;
 import static io.aklivity.zilla.runtime.engine.EngineConfiguration.ENGINE_WORKER_CAPACITY;
+import static io.aklivity.zilla.runtime.engine.namespace.NamespacedId.NO_NAMESPACED_ID;
 import static java.nio.file.FileVisitOption.FOLLOW_LINKS;
 import static java.nio.file.Files.exists;
 import static java.util.Collections.synchronizedList;
@@ -59,6 +60,7 @@ import io.aklivity.zilla.runtime.engine.EngineBuilder;
 import io.aklivity.zilla.runtime.engine.EngineConfiguration;
 import io.aklivity.zilla.runtime.engine.binding.Binding;
 import io.aklivity.zilla.runtime.engine.ext.EngineExtContext;
+import io.aklivity.zilla.runtime.engine.internal.metrics.EngineWorkersUtilizationMetric;
 import io.aklivity.zilla.runtime.engine.test.annotation.Configuration;
 import io.aklivity.zilla.runtime.engine.test.annotation.Configure;
 
@@ -67,10 +69,11 @@ public final class EngineRule implements TestRule
     // needed by test annotations
     public static final String ENGINE_BUFFER_POOL_CAPACITY_NAME = "zilla.engine.buffer.pool.capacity";
     public static final String ENGINE_BUFFER_SLOT_CAPACITY_NAME = "zilla.engine.buffer.slot.capacity";
-    public static final String ENGINE_CONFIG_URL_NAME = "zilla.engine.config.url";
     public static final String ENGINE_CACERTS_STORE_TYPE_NAME = "zilla.engine.cacerts.store.type";
     public static final String ENGINE_CACERTS_STORE_NAME = "zilla.engine.cacerts.store";
     public static final String ENGINE_CACERTS_STORE_PASS_NAME = "zilla.engine.cacerts.store.pass";
+    public static final String ENGINE_CONFIG_URL_NAME = "zilla.engine.config.url";
+    public static final String ENGINE_WORKER_CAPACITY_NAME = "zilla.engine.worker.capacity";
 
     private static final long EXTERNAL_AFFINITY_MASK = 1L << (Long.SIZE - 1);
     private static final Pattern DATA_FILENAME_PATTERN = Pattern.compile("data\\d+");
@@ -83,6 +86,7 @@ public final class EngineRule implements TestRule
     private EngineConfiguration configuration;
     private String configRoot;
     private Predicate<String> exceptions;
+    private boolean interruptible;
     private boolean clean;
 
     public EngineRule()
@@ -90,6 +94,7 @@ public final class EngineRule implements TestRule
         this.builder = Engine.builder();
         this.properties = new Properties();
         this.exceptions = m -> false;
+        this.interruptible = true;
 
         configure(ENGINE_DRAIN_ON_CLOSE, true);
         configure(ENGINE_SYNTHETIC_ABORT, true);
@@ -153,6 +158,13 @@ public final class EngineRule implements TestRule
         Predicate<String> exceptions)
     {
         this.exceptions = exceptions;
+        return this;
+    }
+
+    public EngineRule interruptible(
+        boolean interruptible)
+    {
+        this.interruptible = interruptible;
         return this;
     }
 
@@ -244,6 +256,11 @@ public final class EngineRule implements TestRule
         return engine.context().counterWriter(namespace, binding, metric, core);
     }
 
+    public LongSupplier utilization()
+    {
+        return gauge(NO_NAMESPACED_ID, supplyLabelId(EngineWorkersUtilizationMetric.NAME));
+    }
+
     public int supplyLabelId(
         String label)
     {
@@ -323,9 +340,12 @@ public final class EngineRule implements TestRule
                 final List<Throwable> errors = synchronizedList(new ArrayList<>());
                 final ErrorHandler errorHandler = ex ->
                 {
-                    ex.printStackTrace();
                     errors.add(ex);
-                    baseThread.interrupt();
+
+                    if (interruptible)
+                    {
+                        baseThread.interrupt();
+                    }
                 };
 
                 FileSystem fs = null;