+++ runtime/guard-jwt/src/main/java/io/aklivity/zilla/runtime/guard/jwt/config/JwtOptionsConfig.java
@@ -27,6 +27,7 @@ public class JwtOptionsConfig extends OptionsConfig
 {
     public final String issuer;
     public final String audience;
+    public final String roles;
     public final List<JwtKeyConfig> keys;
     public final Optional<Duration> challenge;
     public final String identity;
@@ -46,6 +47,7 @@ public class JwtOptionsConfig extends OptionsConfig
     JwtOptionsConfig(
         String issuer,
         String audience,
+        String roles,
         List<JwtKeyConfig> keys,
         Duration challenge,
         String identity,
@@ -53,6 +55,7 @@ public class JwtOptionsConfig extends OptionsConfig
     {
         this.issuer = issuer;
         this.audience = audience;
+        this.roles = roles;
         this.keys = keys;
         this.challenge = ofNullable(challenge);
         this.identity = identity;