+++ runtime/vault-filesystem/src/main/java/io/aklivity/zilla/runtime/vault/filesystem/internal/FileSystemContext.java
@@ -15,24 +15,29 @@
  */
 package io.aklivity.zilla.runtime.vault.filesystem.internal;
 
+import static io.aklivity.zilla.runtime.engine.EngineConfiguration.ENGINE_CERTIFICATE_REVOCATION_STRATEGY;
+
 import java.nio.file.Path;
 import java.util.function.Function;
 
 import io.aklivity.zilla.runtime.engine.Configuration;
 import io.aklivity.zilla.runtime.engine.EngineContext;
 import io.aklivity.zilla.runtime.engine.config.VaultConfig;
+import io.aklivity.zilla.runtime.engine.security.RevocationStrategy;
 import io.aklivity.zilla.runtime.engine.vault.VaultContext;
 import io.aklivity.zilla.runtime.vault.filesystem.config.FileSystemOptionsConfig;
 
 final class FileSystemContext implements VaultContext
 {
     private final Function<String, Path> resolvePath;
+    private final RevocationStrategy revocation;
 
     FileSystemContext(
         Configuration config,
         EngineContext context)
     {
         this.resolvePath = context::resolvePath;
+        this.revocation = ENGINE_CERTIFICATE_REVOCATION_STRATEGY.get(config);
     }
 
     @Override
@@ -40,7 +45,7 @@ final class FileSystemContext implements VaultContext
         VaultConfig vault)
     {
         FileSystemOptionsConfig options = (FileSystemOptionsConfig) vault.options;
-        return new FileSystemVaultHandler(options, resolvePath);
+        return new FileSystemVaultHandler(options, resolvePath, revocation);
     }
 
     @Override